python/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
PyYAML>=6.0
markdown>=3.4.0
python-docx>=0.8.11
